auth/requirements.txt
alembic==1.7.7
apispec==5.2.1
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
flasgger==0.9.5
Flask==2.1.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
flask-paginate==2022.1.8
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.5.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mistune==2.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
redis==4.2.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
